Update config_template from OSA repository
The plugin is currently broken with Ansible 2.5 because it uses an
import which was deprecated in Ansible 2.2 and removed in Ansible 2.5.

The upstream for this action plugin is the OSA openstack-ansible-plugins
repository. This repository contains a fix for Ansible 2.5
compatibility.

Upstream HEAD commit a7d1e6375e7d21dacf1ff71f8bd38cc64508e2bf.