Change parameter name for Ansible 2.7 compatibility
The parameter used by the config_template plugin was renamed from
original_basename to _original_basename to better reflect that this is
an internal parameter not to be used directly by users. This commit
changes the plugin to use _original_basename. This is probably OK
because it's in a plugin and not a direct usage.